requirements.txt
spacy
pdfplumber
python-docx
